Add support for DOMAIN-ROUTE processing
Add support for DOMAIN-ROUTE as an option to be passed to the script which will specify the listed domain to be routed using systemd-resolved via this VPN link using the DNS servers specified. This is not the same as DOMAIN or DOMAIN-SEARCH, as both of those options will allow systemd-resolved to append them to any bare name for searching (as in keeping with update-resolv-conf and update-systemd-network), and to some extent, Windows and macOS as well.