Handle the privlege drop case when down() is called
If down() is called when the user/group options have been set, the OpenVPN client (and hence the script that is forked from it) no longer have the required privileges to call the RevertLink function over DBus. Rather than have the system throw a big error, make a note in the logs and exit gracefully.